+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pmaltese/prisma-schema-dsl-types
 
+## 1.1.0
+
+### Minor Changes
+
+- Implemented: generator preview features, model indexes and fullText indexes
+
 ## 1.0.10
 
 ### Patch Changes